https://ibax.io/feature/architecture
High Performance / High TPS Blockchain

The popular blockchain represented by Ethereum has a low transaction processing capacity, which seriously hinders the development of blockchain. A high performance/high concurrency blockchain will bring leaps and bounds to the whole blockchain industry. The IBAX network is as fast as 4 seconds a block and transaction processing up to 3000 TPS. With the improvement of hardware, the TPS can theoretically reach 10000 times in the case of unsliced.